Correcciones backend, frontend y base de datos

+++ 03_backend/proyect/src/main/java/com/backend/proyect/dto/devoluciones/DevolucionesCambiosRequest.java
@@ -8,7 +8,8 @@ public class DevolucionesCambiosRequest {
     private String fechaSolicitud;
     private String fechaRespuesta;
 
-    private Long idUsuario;
+
+    private Integer idUsuario;
 
     // Getters y Setters
     public String getMotivo() {
@@ -51,11 +52,9 @@ public class DevolucionesCambiosRequest {
         this.fechaRespuesta = fechaRespuesta;
     }
 
-    public Long getIdUsuario() {
-        return idUsuario;
-    }
+    public Integer getIdUsuario() { return idUsuario; }
 
-    public void setIdUsuario(Long idUsuario) {
+    public void setIdUsuario(Integer idUsuario) {
         this.idUsuario = idUsuario;
     }
